refactor(movies): assets folder name + unit tests
