+++ package.json
@@ -7,7 +7,8 @@
     "heroku-prebuild": "cp ./src/config/env.ts.dist ./src/config/env.ts",
     "heroku-postbuild": "yarn db:seed",
     "postinstall": "if-env IS_HEROKU=true && yarn build || (yarn build && yarn docker && yarn test)",
-    "build": "tsc && yarn doc",
+    "clear": "rimraf dist",
+    "build": "yarn clear && tsc && yarn doc",
     "start": "yarn start:prod",
     "start:dev": "ts-node --project tsconfig.json src/index.ts",
     "start:prod": "node dist/index.js",
