+++ src/tests/movieActor.mock.ts
@@ -1,5 +1,5 @@
 import * as faker from 'faker';
-import { MovieActor } from '../api/movie/model/movie.model';
+import { MovieActor } from '../api/movies/model/movies.model';
 
 export const mockMovieActor = (): MovieActor => ({
   imdbId: faker.random.uuid(),
